requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
rich
psutil
